server/requirements.txt
aenum==3.1.15
aiohttp==3.8.4
aiosignal==1.3.1
allure-pytest==2.13.5
allure-python-commons==2.13.5
altgraph==0.17.4
anyio==3.6.2
appdirs==1.4.4
argcomplete==3.0.5
assertpy==1.1
astroid==2.15.8
asttokens==2.2.1
async-timeout==4.0.2
attrs==24.2.0
Babel==2.8.0
backcall==0.2.0
bcc==0.18.0
beautifulsoup4==4.12.3
black==23.1.0
blinker==1.8.2
Brlapi==0.8.3
Brotli==1.1.0
cached-property==1.5.2
cachelib==0.9.0
cachetools==5.3.3
certifi==2024.8.30
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.1.0
chrome-gnome-shell==0.0.0
click==8.1.7
codespell==2.1.0
colorama==0.4.4
command-not-found==0.3
commonmark==0.9.1
ConfigArgParse==1.7
cryptography==43.0.1
cssselect==1.2.0
cupshelpers==1.0
curlify==2.2.1
dataclasses-json==0.5.7
dbus-python==1.2.18
decorator==5.1.1
defer==1.0.6
delayed-assert==0.3.6
Deprecated==1.2.13
dill==0.3.9
distro==1.7.0
dnspython==2.7.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
docstring_parser==0.16
email_validator==2.2.0
evdev==1.4.0
exceptiongroup==1.2.2
execnet==2.1.1
executing==1.2.0
exrex==0.11.0
Faker==30.3.0
fast-map==0.1.0
fastapi==0.115.2
fastapi-cli==0.0.5
fastapi-utils==0.7.0
feedparser==6.0.8
flake8==6.0.0
Flask==3.0.3
Flask-BasicAuth==0.2.0
Flask-Cors==5.0.0
frozenlist==1.3.3
gevent==21.12.0
geventhttpclient==2.3.1
git-filter-repo==2.38.0
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.135.0
google-auth==2.31.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.58.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.3
google-cloud-storage==2.17.0
google-cloud-texttospeech==2.16.3
google-crc32c==1.5.0
google-generativeai==0.7.1
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
gpg==1.16.0
greenlet==1.1.2
grpc-google-iam-v1==0.13.1
grpcio==1.64.1
grpcio-status==1.62.2
gunicorn==23.0.0
gyp==0.1
h11==0.14.0
halo==0.0.31
hidpidaemon==18.4.6
howdoi==2.0.20
html2text==2020.1.16
httpcore==0.16.3
httplib2==0.20.2
httptools==0.6.1
httpx==0.23.3
hubspace-ng==0.0.4
idna==3.3
imageio==2.34.2
imageio-ffmpeg==0.5.1
importlib-metadata==4.6.4
inflect==7.3.1
influxdb-client==1.46.0
iniconfig==2.0.0
isort==5.13.2
itsdangerous==2.2.0
jedi==0.18.2
jeepney==0.7.1
Jinja2==3.1.4
jsonpath-ng==1.7.0
jsonschema==3.2.0
keep==2.10.1
kernelstub==3.1.4
keyring==23.5.0
langchain==0.0.138
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy-object-proxy==1.10.0
llama-index==0.5.13.post1
locust==2.12.1
log-symbols==0.0.14
louis==3.20.0
lxml==4.9.2
macaroonbakery==1.3.1
MarkupSafe==3.0.1
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib-inline==0.1.6
mccabe==0.7.0
more-itertools==8.10.0
moviepy==1.0.3
msgpack==1.0.4
multidict==6.0.4
mypy-extensions==1.0.0
mysql-connector-python==8.0.32
mysqlclient==1.4.6
netaddr==0.8.0
numpy==1.24.2
oauthlib==3.2.0
openai==0.27.4
openapi-schema-pydantic==1.2.4
outcome==1.3.0.post0
packaging==23.0
pandas==1.5.0
parso==0.8.3
pathspec==0.12.1
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.4.0
pipx==1.2.0
pkce==1.0.3
platformdirs==4.3.6
please-cli==0.2.1
pluggy==1.5.0
ply==3.11
pop-transition==1.1.2
proglog==0.1.10
prometheus_client==0.21.0
prompt-toolkit==3.0.38
proto-plus==1.24.0
protobuf==4.25.3
psutil==5.9.0
psycogreen==1.0.2
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyByteBuffer==1.0.5
pycairo==1.20.1
pycodestyle==2.10.0
pycparser==2.21
pycups==2.0.1
pydantic==1.10.7
pydbus==0.6.0
pyflakes==3.0.1
PyGithub==1.57
Pygments==2.11.2
PyGObject==3.42.1
pyinotify==0.9.6
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
PyJWT==2.6.0
pylint==2.17.0
pymacaroons==0.13.0
PyMySQL==1.0.3
PyNaCl==1.5.0
pynvim==0.4.3
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-sip==12.9.1
PyQtWebEngine==5.15.5
pyquery==1.4.3
pyRFC3339==1.1
pyrsistent==0.18.1
PySocks==1.7.1
pytest==8.3.3
pytest-json-report==1.5.0
pytest-metadata==3.1.1
pytest-order==1.3.0
pytest-reportportal==5.1.5
pytest-rerunfailures==14.0
pytest-xdist==3.6.1
python-apt==2.4.0+ubuntu3
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
python-gnupg==0.4.8
python-Levenshtein==0.12.2
python-multipart==0.0.12
python-xlib==0.29
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==26.2.0
qutebrowser==2.5.0
reactivex==4.0.4
regex==2023.3.23
repolib==2.2.1
repoman==1.4.0
reportportal-client==5.2.5
requests==2.28.2
rfc3986==1.5.0
rich==12.5.1
roundrobin==0.0.4
rsa==4.9
SecretStorage==3.3.1
selenium==4.21.0
sessioninstaller==0.0.0
sgmllib3k==1.0.0
shapely==2.0.4
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.6
spinners==0.0.24
SQLAlchemy==1.4.47
ssh-import-id==5.11
stack-data==0.6.2
starlette==0.39.2
system76driver==20.4.95
systemd-python==234
tenacity==8.2.2
termcolor==2.2.0
terminaltables==3.1.10
texttable==1.6.4
tiktoken==0.3.3
toml==0.10.2
tomli==2.0.2
tomlkit==0.13.2
tqdm==4.65.0
traitlets==5.9.0
trio==0.26.2
trio-websocket==0.11.1
typeguard==4.3.0
typer==0.12.5
typing-inspect==0.8.0
typing_extensions==4.12.2
tzdata==2023.3
ubuntu-drivers-common==0.0.0
ufw==0.36.1
ulauncher==5.14.0
uritemplate==4.1.1
urllib3==1.26.5
userpath==1.8.0
uvicorn==0.31.1
uvloop==0.20.0
wadllib==1.3.6
watchfiles==0.24.0
wcwidth==0.2.6
webdriver-manager==4.0.2
websocket-client==1.4.2
websockets==12.0
Werkzeug==3.0.4
wrapt==1.14.1
wsproto==1.2.0
xdg==5
xkit==0.0.0
yarl==1.8.2
zipp==1.0.0
zope.event==5.0
zope.interface==7.1.0
